# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Process-Sync-Api",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/process-definitions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "list process-definitions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "process-definitions"
                ],
                "summary": "list process-definitions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "default 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "default id.asc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma seperated list of network-ids used to filter",
                        "name": "network_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ProcessDefinition"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/process-definitions/{networkId}/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get process-definition",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "process-definitions"
                ],
                "summary": "get process-definition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "process definition id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ProcessDefinition"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/process-instances": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "list process-instances",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "process-instance"
                ],
                "summary": "list process-instances",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "default 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "default id.asc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma seperated list of network-ids used to filter",
                        "name": "network_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ProcessInstance"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/process-instances/{networkId}/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get process-instances",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "process-instance"
                ],
                "summary": "get process-instances",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "instance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HistoricProcessInstance"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get process-instances",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "process-instance"
                ],
                "summary": "get process-instances",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "instance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.HistoricProcessInstance": {
            "type": "object",
            "properties": {
                "businessKey": {
                    "type": "string"
                },
                "caseInstanceId": {
                    "type": "string"
                },
                "deleteReason": {
                    "type": "string"
                },
                "durationInMillis": {
                    "type": "number"
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_placeholder": {
                    "type": "boolean"
                },
                "marked_as_missing": {
                    "type": "boolean"
                },
                "marked_for_delete": {
                    "type": "boolean"
                },
                "network_id": {
                    "type": "string"
                },
                "processDefinitionId": {
                    "type": "string"
                },
                "processDefinitionKey": {
                    "type": "string"
                },
                "processDefinitionName": {
                    "type": "string"
                },
                "processDefinitionVersion": {
                    "type": "number"
                },
                "startActivityId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "startUserId": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "superCaseInstanceId": {
                    "type": "string"
                },
                "superProcessInstanceId": {
                    "type": "string"
                },
                "sync_date": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                }
            }
        },
...
        "model.ProcessDefinition": {
            "type": "object",
            "properties": {
                "Version": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "deploymentId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "diagram": {
                    "type": "string"
                },
                "historyTimeToLive": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_placeholder": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "marked_as_missing": {
                    "type": "boolean"
                },
                "marked_for_delete": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "network_id": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                },
                "suspended": {
                    "type": "boolean"
                },
                "sync_date": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                },
                "versionTag": {
                    "type": "string"
                }
            }
        },
        "model.ProcessInstance": {
            "type": "object",
            "properties": {
                "businessKey": {
                    "type": "string"
                },
                "caseInstanceId": {
                    "type": "string"
                },
                "definitionId": {
                    "type": "string"
                },
                "ended": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_placeholder": {
                    "type": "boolean"
                },
                "marked_as_missing": {
                    "type": "boolean"
                },
                "marked_for_delete": {
                    "type": "boolean"
                },
                "network_id": {
                    "type": "string"
                },
                "suspended": {
                    "type": "boolean"
                },
                "sync_date": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}